RPM: Use Python 3 to fix pipe encoding
"setup.py" replaces the shebang line in "leo" with the interpreter that
was used to call "setup.py". Call "setup.py" with python3 so that the
result is consistent with the shebang in git.

For the user, this fixes encoding problems when using pipes, i.e. fixes:

  $ leo Fürst | grep Finsternis
  Prince of Darkness              Fürst der Finsternis

(Using Python 2.x, this led to 'ascii' codec error messages.)