fix errors printing inflection tables
Fix the the `_pairwise` function, which would always throw a 
`StopIteration` error instead of exiting gracefully at the end of an
iterator. The new code also requires importing `izip` on Python 2.
Fixes #18.

Running `leo -i gelb` gave this error:

    'NoneType' object has no attribute 'get_text'

This was because `Gelb` does not have a separable ending. Account for
this by setting the variable to `''` by default.